Lift EmojiHistoryContext
